refactor: fenwick_tree.hpp を lib/fenwick/ から lib/data_structure/ へ移動
lib/fenwick_tree_raq.hpp 統合後 lib/fenwick/ が単一ファイルのディレクトリに
なったため、他の汎用データ構造（swag、range_set 等）と同じ lib/data_structure/
にまとめてディレクトリを削除。
