+++ test/yosupo/data_structure/point_add_range_sum.test.cpp
@@ -1,7 +1,7 @@
 // competitive-verifier: PROBLEM https://judge.yosupo.jp/problem/point_add_range_sum
 #include <cstdint>
 #include <iostream>
-#include "fenwick/fenwick_tree.hpp"
+#include "data_structure/fenwick_tree.hpp"
 #include "template/sonic.hpp"
 
 int main(void) {
