+++ test/yosupo/data_structure/static_range_inversions.test.cpp
@@ -5,7 +5,7 @@
 #include <vector>
 #include "algorithm/compress.hpp"
 #include "algorithm/mo.hpp"
-#include "fenwick/fenwick_tree.hpp"
+#include "data_structure/fenwick_tree.hpp"
 
 int main(void) {
     int n, q;
