+++ test/yosupo/tree/vertex_add_path_sum.test.cpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 #include <iostream>
 #include <vector>
-#include "fenwick/fenwick_tree.hpp"
+#include "data_structure/fenwick_tree.hpp"
 #include "graph/edge_input.hpp"
 #include "tree/hld.hpp"
