+++ test/yosupo/tree/vertex_add_subtree_sum.2.test.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <utility>
 #include <vector>
-#include "fenwick/fenwick_tree.hpp"
+#include "data_structure/fenwick_tree.hpp"
 #include "graph/edge_input.hpp"
 #include "tree/dsu_on_tree.hpp"
 #include "union_find/union_find.hpp"
